[Improve] interface로 분리
